Added empty field checking, and enforced unique usernames
