+++ app/src/main/AndroidManifest.xml
@@ -12,10 +12,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.CMPUT301Project"
         tools:targetApi="31">
-        <activity android:name=".SignUpActivity"
+        <activity android:name=".activities.SignUpActivity"
             android:exported="true"/>
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
